apple-gmux: Handle external DP port properly on 2011+ MacBook Pros
On MacBook Pros introduced 2011 and onward, the external DP port is
a combined DP/Thunderbolt port that is no longer fully switchable
between GPUs, it can only be driven by the discrete GPU. When a display
is plugged in while the discrete GPU is asleep, wake it up. The HPD pin
is connected to gmux which sends an interrupt on hotplug.

Oddly enough, while the full port is not switchable on these machines,
AUX still is. Since the integrated GPU is missing the main link,
external displays appear to it as phantoms which fail to link-train.
Force the AUX channel to the discrete GPU to avoid any confusion.

Document the switching policy implemented by this commit.

Actually we would also need to wake the Thunderbolt controller on HPD
but its driver currently doesn't support runtime pm.

Signed-off-by: Lukas Wunner <lukas@wunner.de>