vga_switcheroo: Cope with handler and DRM driver in separate modules
On non-Mac dual GPU laptops, the vga_switcheroo handler (v1_dsm, optimus,
atpx) is in the same module as the DRM driver for the discrete GPU.
On Macs however, the handler (apple-gmux) is in a *separate* module.
Neither the DRM drivers nor vga_switcheroo acquire a reference on the
handler module, so the module may unload at any given time.

To cope with this:

- Reinstate power to the GPUs when the handler unregisters.
  (Currently no handler is capable of cutting power to the integrated
  GPU, still reinstate power for that one as well in case such a handler
  ever exists.)

- If driver power control is used, avoid an oops when dereferencing
  vgasr_priv.handler in vga_switcheroo_runtime_suspend() and _resume().

Signed-off-by: Lukas Wunner <lukas@wunner.de>