vga_switcheroo: Don't oops on runtime pm if no handler is registered
Nouveau, radeon and amdgpu call vga_switcheroo_init_domain_pm_ops()
if a handler capable of cutting power to the discrete GPU is detected
in the system. This works fine with nouveau's v1_dsm and atpx since
the handler and the DRM driver are in the same module. With apple-gmux,
the handler is in a separate module which may be unloaded at any given
time. (The DRM drivers do not acquire a reference to the apple-gmux
module.)

If the handler module is not loaded, vga_switcheroo_runtime_suspend()
and vga_switcheroo_runtime_resume() will oops because they dereference
vgasr_priv.handler. Make those functions a no-op if there's no handler.

Also, when the handler unregisters prematurely, be sure to reinstate
power to the GPUs. (Currently no handler is capable of cutting power
to the integrated GPU, still reinstate power for that one as well
in case such a handler ever exists.)

Signed-off-by: Lukas Wunner <lukas@wunner.de>