drm/dp + drm/edid: Use vga_switcheroo active client for AUX proxying
The retina MacBook Pro uses an eDP panel and a gmux controller to switch
the panel between its two GPUs. Unfortunately it cannot switch the AUX
channel separately from the main link, making it impossible for the
inactive GPU to read the EDID and perform link training unless we
switch over the entire panel, causing additional flickering while
loading the DRM drivers.

But we can emulate AUX switching in software by using the active client
as a proxy to talk to the panel.

When reading from the DPCD or EDID, retrieve the active client's struct
drm_dp_aux and use it in lieu of the client's own struct drm_dp_aux.

When writing to the DPCD, compare what's to be written with the current
DPCD content and skip the write if it's identical. If not, the write
will simply fail. This should make it possible to complete link training
if both drivers use the same lane count, bandwidth, etc.

To avoid using proxying with external DP-connected displays, it is
constrained to eDP connectors for DPCD reads/writes. For EDID reads,
the onus is on the DRM drivers to only call drm_get_edid_switcheroo()
for eDP connectors.

Takes inspiration from patches by Matthew Garrett <mjg59@srcf.ucam.org>
who used stashing instead of proxying:
http://www.codon.org.uk/~mjg59/tmp/retina_patches/0025-i915-Use-vga_switcheroo-to-obtain-and-stash-EDID-DPC.patch
http://www.codon.org.uk/~mjg59/tmp/retina_patches/0026-nouveau-Use-vga_switcheroo-to-obtain-and-stash-EDID-.patch

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=88861
Signed-off-by: Lukas Wunner <lukas@wunner.de>