drm/i915: Use vga_switcheroo active client for AUX proxying
The retina MacBook Pro uses an eDP panel and a gmux controller to switch
the panel between its two GPUs. The panel mode in VBIOS is notoriously
bogus on these machines and some models have no VBIOS at all. Switching
the AUX channel separately from the main link is not supported by the
hardware, but we can emulate AUX switching in software by using the
active GPU as a proxy to talk to the panel.

Defer probing if the vga_switcheroo handler flags indicate that AUX
proxying is needed, we're the inactive GPU and the driver for the active
GPU has not initialized yet.

Use drm_get_edid_switcheroo() in lieu of drm_get_edid() on eDP.
This allows us to retrieve the EDID if the panel is currently muxed
to the discrete GPU.

List of retina MBPs with dual GPUs, one of them Intel:
    [MBP 10,1 2012  intel IVB + nvidia GK107      retina  15"]
    [MBP 11,3 2013  intel HSW + nvidia GK107      retina  15"]
    [MBP 11,5 2015  intel HSW + amd cape verde    retina  15"]

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=88861
Signed-off-by: Lukas Wunner <lukas@wunner.de>