drm/nouveau/i2c: Use vga_switcheroo active client for AUX proxying
The retina MacBook Pro uses an eDP panel and a gmux controller to switch
the panel between its two GPUs. The panel mode in VBIOS is notoriously
bogus on these machines and some models have no VBIOS at all. Switching
the AUX channel separately from the main link is not supported by the
hardware, but we can emulate AUX switching in software by using the
active GPU as a proxy to talk to the panel.

AUX proxying via vga_switcheroo works by replacing the inactive GPU's
struct drm_dp_aux by that of the active GPU. However, as regards usage
of struct drm_dp_aux, nouveau is the odd man out: A struct drm_dp_aux
is defined as part of struct nouveau_connector but never used. Instead,
the AUX channel is accessed directly with nvkm_rdaux() and nvkm_wraux()
even in the DRM part of the driver.

To enable proxying in nouveau, inject a pointer to the struct drm_dp_aux
from the DRM part of the driver into struct nvkm_i2c_aux. Modify
nvkm_rdaux() to try drm_dp_dpcd_read() first. If that fails, fall back
to accessing the AUX channel directly. Likewise, modify nvkm_wraux() to
compare what's to be written with the current DPCD content and skip
the write if it's identical. Constrain to platforms whose vga_switcheroo
handler flags explicitly indicate that proxying is needed. Enclose
in #if to keep the nvkm part of the driver free of DRM symbols.

When the driver first initializes the output with nvkm_output_dp_init(),
the pointer to the struct drm_dp_aux is not yet injected into the struct
nvkm_i2c_port, in fact it hasn't even been created yet. Thus, if the
panel is not switched to the Nvidia GPU, the dpcd attribute of struct
nvkm_output_dp can't be filled and the link doesn't get trained. Make up
for this by checking the link training status in nouveau_dp_detect() and
calling nvkm_output_dp_enable() if the link hasn't been trained yet.
(As a side effect, this fixes split brain issues in nouveau since the
nvkm part of the driver only attempts link training on driver load
and on HPD, whereas the DRM part regularly polls the outputs, so the
DRM part may detect a display where the nvkm part doesn't.)

Defer probing if the vga_switcheroo handler flags indicate that AUX
proxying is needed, we're the inactive GPU and the driver for the active
GPU has not yet initialized.

Finally, use drm_get_edid_switcheroo() in lieu of drm_get_edid() on eDP.

List of retina MBPs with dual GPUs, one of them Intel:
    [MBP 10,1 2012  intel IVB + nvidia GK107      retina  15"]
    [MBP 11,3 2013  intel HSW + nvidia GK107      retina  15"]

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=88861
Signed-off-by: Lukas Wunner <lukas@wunner.de>