thunderbolt: Support 1st gen Light Ridge controller
Built into:
iMac12,1       2011  21.5"
iMac12,2       2011  27"
Macmini5,1     2011  i5 2.3 GHz
Macmini5,2     2011  i5 2.5 GHz
Macmini5,3     2011  i7 2.0 GHz
MacBookPro8,1  2011  13"
MacBookPro8,2  2011  15"
MacBookPro8,3  2011  17"
MacBookPro9,1  2012  15"
MacBookPro9,2  2012  13"

The chip supports a maximum hop_count of 32, unlike its successors
which support only 12. Fixup the calculation of reg and bit in
ring_interrupt_active() to cope with values >= 32.

Another peculiarity is that the chip supports a maximum of 13 ports
whereas its successors support 12. However the additional port (#5)
seems to be unusable as reading its TB_CFG_PORT config space results
in TB_CFG_ERROR_INVALID_CONFIG_SPACE. Add a quirk to mark the port
disabled, assuming that this is necessary on all models using this chip.

As with 2nd gen Cactus Ridge controllers, the uid in drom is not unique,
it's 0x1000000000000 on every chip (with matching crc8). The OS X driver
uses a different uid, it is unclear where it's coming from. Possibly
the EFI firmware.

Apple uses combined DP/Thunderbolt ports, requiring a mechanism to
detect which type of device is connected. The detection mechanism for
this 1st gen Thunderbolt chip is awfully complicated and consists of
a microcontroller (NXP LPC1112A), DP redriver (Parade PS8301) and
mux (NXP CBTL04DP081), the former two are connected to the Thunderbolt
controller with I2C. Thankfully the interaction between these chips
and the detection of the type of device seems to work automatically,
without OS support. However OS support might be necessary to put the
chips in power saving mode if they're not used.

These additional chips are missing on models using the 2nd gen Cactus
Ridge controller, the ability to differentiate between DP/Thunderbolt
devices was apparently integrated into the controller. If a DP device
is connected, the Thunderbolt controller routes the signals to one of
its DP sink ports. On the MacBook Pro, this frees up one of the four
ports on the GPU, which allowed Apple to add HDMI to the retina.
(Pre-retinas used the ports for 1x LVDS, 1x external DP, 2x Thunderbolt.
Retinas use them for 1x embedded DP, 1x external HDMI, 2x Thunderbolt.)

Data sheets:
http://www.nxp.com/documents/data_sheet/LPC111X.pdf
http://www.nxp.com/documents/data_sheet/CBTL04DP211.pdf
http://www.paradetech.com/wp-content/uploads/2009/08/PS8311-Product-Brief-20090811.pdf

Cc: Andreas Noever <andreas.noever@gmail.com>
Tested-by: Lukas Wunner <lukas@wunner.de> [MacBookPro9,1]
Tested-by: William Brown <william@blackhats.net.au> [MacBookPro8,2]
Signed-off-by: Lukas Wunner <lukas@wunner.de>