thunderbolt: Replace PCI quirk for downstream bridges
When resuming from system sleep, the downstream bridges of a Thunderbolt
controller must not resume before the NHI has reestablished the PCI
tunnels. However the NHI is not a parent of the downstream bridges,
so this requirement is not fulfilled automatically by the PM core and we
have to open code it instead. Up until now we've done this with a PCI
quirk, which has the following disadvantages:

- The code for the NHI resides in drivers/thunderbolt/, whereas the code
  for the downstream bridges resides in drivers/pci/quirks.c, which is
  somewhat surprising and non-obvious in particular for newcomers.

- Whenever support for an additional Thunderbolt controller is added,
  its PCI device ID needs to be amended in both places, which invites
  mistakes. E.g. commit a42fb351ca1f ("thunderbolt: Allow loading of
  module on recent Apple MacBooks with thunderbolt 2 controller")
  relaxed the subvendor and subdevice ID in the NHI code but forgot to
  also change the downstream bridge code.

- Since the PCI quirk cannot keep any state, it has to search for the
  NHI over and over again on every resume.

Overcome these disadvantages by resuming the NHI from the upstream
bridge driver. The upstream bridge *is* a parent of the downstream
bridges, so this is functionally equivalent to the PCI quirk. Also
needs less code.

Cc: Andreas Noever <andreas.noever@gmail.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>