PCI: Whitelist Thunderbolt ports for runtime PM
Currently PCIe ports are only allowed to go to D3hot if the BIOS is
dated 2015+ to avoid potential issues with old chipsets.  However for
Thunderbolt we know that even the oldest controller, Light Ridge (2010),
is able to suspend its ports to D3hot just fine.

Each Thunderbolt controller consumes ~2W even if nothing is plugged in
and runtime suspending its PCIe ports to D3hot allows the controller to
be powered down.   The achievable power saving should be afforded to
machines even if their BIOS predates the 2015 cutoff, so whitelist
these ports.

Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Andreas Noever <andreas.noever@gmail.com>
Cc: Michael Jamet <michael.jamet@intel.com>
Cc: Tomas Winkler <tomas.winkler@intel.com>
Cc: Amir Levy <amir.jer.levy@intel.com>
Reviewed-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>