vga_switcheroo: Rename ->reprobe callback to ->post_switch
On muxed laptops, the runtime pm refcount needs to be updated for both
GPUs when switching the mux: The newly active GPU needs to hold a ref
for the panel CRTC and the previously active GPU needs to release a ref.

We already have a ->reprobe callback which is invoked after the switch
for the newly active GPU. Rename it to ->post_switch and call it for
both GPUs.

Currently only nouveau is actually declaring a ->reprobe callback. All
it does is polling the outputs. To keep this a pure refactoring change
without functional side effects, amend nouveau's callback to only poll
the outputs if the GPU is the newly active one.

Signed-off-by: Lukas Wunner <lukas@wunner.de>